Split out getLoginUrlWithTOSRedirect
