+++ client/lib/gsuite/index.js
@@ -16,6 +16,7 @@ export {
 	isGSuiteOrExtraLicenseProductSlug,
 	isGSuiteProductSlug,
 } from './gsuite-product-slug';
+export { getLoginUrlWithTOSRedirect } from './get-login-url-with-tos-redirect';
 
 /**
  * Determines whether G Suite is allowed for the specified domain.
@@ -81,24 +82,6 @@ export function getGSuiteSupportedDomains( domains ) {
 
 export function getGSuiteMailboxCount( domain ) {
 	return get( domain, 'googleAppsSubscription.totalUserCount', 0 );
-}
-
-/**
- * Creates the Google ToS redirect url given email and domain
- *
- * @param {string} email - email
- * @param {string} domain - domain name
- * @returns {string} - ToS url redirect
- */
-export function getLoginUrlWithTOSRedirect( email, domain ) {
-	return (
-		'https://accounts.google.com/AccountChooser?' +
-		`Email=${ encodeURIComponent( email ) }` +
-		`&service=CPanel` +
-		`&continue=${ encodeURIComponent(
-			`https://admin.google.com/${ domain }/AcceptTermsOfService?continue=https://mail.google.com/mail/u/${ email }`
-		) }`
-	);
 }
 
 /**
